# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Chillist API",
    "description": "Trip planning with shared checklists",
    "version": "1.0.0"
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "def-0": {
        "type": "object",
        "properties": {
          "message": {
            "type": "string"
          },
          "code": {
            "type": "string"
          }
        },
        "required": [
          "message"
        ],
        "title": "ErrorResponse"
      },
...
      "def-54": {
        "type": "object",
        "properties": {
          "foodPreferences": {
            "type": "string",
            "nullable": true,
            "description": "Free-text dietary preferences, e.g. \"vegetarian, no shellfish\""
          },
          "allergies": {
            "type": "string",
            "nullable": true,
            "description": "Free-text allergy list, e.g. \"nuts, gluten, dairy\""
          },
          "defaultEquipment": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "description": "List of equipment item names the user typically brings on trips, e.g. [\"tent\", \"sleeping bag\", \"headlamp\"]"
          }
        },
        "title": "UserPreferences"
      },
      "def-55": {
        "type": "object",
        "properties": {
          "user": {
            "type": "object",
            "properties": {
              "id": {
                "type": "string"
              },
              "email": {
                "type": "string"
              },
              "role": {
                "type": "string"
              }
            },
            "required": [
              "id",
              "email",
              "role"
            ]
          },
          "preferences": {
            "oneOf": [
              {
                "$ref": "#/components/schemas/def-54"
              },
              {
                "type": "null"
              }
            ],
            "description": "App preferences for this user. Null if the user has never saved preferences."
          }
        },
        "required": [
          "user",
          "preferences"
        ],
        "title": "ProfileResponse"
      },
      "def-56": {
        "type": "object",
        "properties": {
          "foodPreferences": {
            "type": "string",
            "nullable": true,
            "description": "Free-text dietary preferences, e.g. \"vegetarian, no shellfish\". Send null to clear."
          },
          "allergies": {
            "type": "string",
            "nullable": true,
            "description": "Free-text allergy list, e.g. \"nuts, gluten, dairy\". Send null to clear."
          },
          "defaultEquipment": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true,
            "description": "List of equipment item names the user typically brings. Send null to clear."
          }
        },
        "title": "UpdateProfileBody"
      },
      "def-57": {
        "type": "object",
        "properties": {
          "preferences": {
            "$ref": "#/components/schemas/def-54"
          }
        },
        "required": [
          "preferences"
        ],
        "title": "UpdateProfileResponse"
      },
...
    }
  },
  "paths": {
...
    "/auth/me": {
      "get": {
        "summary": "Get current user from JWT",
        "tags": [
          "auth"
        ],
        "description": "Returns the authenticated user identity extracted from the JWT. Returns 401 if no valid JWT is provided.",
        "responses": {
          "200": {
            "description": "Authenticated user identity",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Authenticated user identity",
                  "type": "object",
                  "properties": {
                    "user": {
                      "type": "object",
                      "properties": {
                        "id": {
                          "type": "string"
                        },
                        "email": {
                          "type": "string"
                        },
                        "role": {
                          "type": "string"
                        }
                      },
                      "required": [
                        "id",
                        "email",
                        "role"
                      ]
                    }
                  },
                  "required": [
                    "user"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "JWT token missing or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "description": "JWT token missing or invalid",
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
    "/auth/profile": {
      "get": {
        "summary": "Get current user profile",
        "tags": [
          "auth"
        ],
        "description": "Returns user identity from JWT and app preferences from the database. Returns 401 if no valid JWT is provided.",
        "responses": {
          "200": {
            "description": "User identity and preferences",
            "content": {
              "application/json": {
                "schema": {
                  "description": "User identity and preferences",
                  "$ref": "#/components/schemas/def-55"
                }
              }
            }
          },
          "401": {
            "description": "JWT token missing or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "description": "JWT token missing or invalid",
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      },
      "patch": {
        "summary": "Update user preferences",
        "tags": [
          "auth"
        ],
        "description": "Creates or updates the authenticated user's app preferences. Returns 401 if no valid JWT is provided.",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/def-56"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated preferences",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Updated preferences",
                  "$ref": "#/components/schemas/def-57"
                }
              }
            }
          },
          "401": {
            "description": "JWT token missing or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "description": "JWT token missing or invalid",
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
    "/auth/sync-profile": {
      "post": {
        "summary": "Sync JWT identity to all participant records",
        "tags": [
          "auth"
        ],
        "description": "Updates all participant records linked to the authenticated user with identity fields from the JWT (name, email, phone, avatar). Call this after updating the user profile in Supabase so all plans reflect the latest data.",
        "responses": {
          "200": {
            "description": "Number of participant records synced",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Number of participant records synced",
                  "type": "object",
                  "properties": {
                    "synced": {
                      "type": "integer"
                    }
                  },
                  "required": [
                    "synced"
                  ]
                }
              }
            }
          },
          "401": {
            "description": "JWT token missing or invalid",
            "content": {
              "application/json": {
                "schema": {
                  "description": "JWT token missing or invalid",
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          },
          "500": {
            "description": "Internal server error",
            "content": {
              "application/json": {
                "schema": {
                  "description": "Internal server error",
                  "$ref": "#/components/schemas/def-0"
                }
              }
            }
          }
        }
      }
    },
...
  },
...
}